+++ crispy_forms/templatetags/crispy_forms_tags.py
@@ -1,9 +1,10 @@
+from functools import lru_cache
+
 from django import template
 from django.conf import settings
 from django.forms.formsets import BaseFormSet
 from django.template.loader import get_template
 
-from crispy_forms.compatibility import lru_cache
 from crispy_forms.helper import FormHelper
 from crispy_forms.utils import TEMPLATE_PACK, get_template_pack