+++ crispy_forms/tests/test_layout_objects.py
@@ -1,3 +1,5 @@
+import re
+
 from django import forms
 from django.template import Context, Template
 from django.test.html import parse_html
@@ -23,7 +25,7 @@ from crispy_forms.layout import HTML, Field, Layout, MultiWidgetField
 from crispy_forms.tests.utils import contains_partial
 from crispy_forms.utils import render_crispy_form
 
-from .conftest import only_bootstrap
+from .conftest import only_bootstrap, only_bootstrap4
 from .forms import CheckboxesSampleForm, SampleForm
 
 
@@ -223,6 +225,22 @@ class TestBootstrapLayoutObjects:
             assert 'class="form-control-lg' in html
             assert contains_partial(html, '<span class="input-group-text"/>')
 
+    @only_bootstrap4
+    def test_prepended_appended_text_in_select(self, settings):
+        test_form = SampleForm()
+        test_form.fields["select"] = forms.ChoiceField(
+            label="Select field", choices=[(1, "Choice 1"), (2, "Choice 2")]
+        )
+
+        test_form.helper = FormHelper()
+        test_form.helper.layout = Layout(
+            AppendedText("select", "USD"),
+        )
+        html = render_crispy_form(test_form)
+
+        assert html.count("custom-select") == 1
+        assert html.count("USD") == 1
+
     def test_inline_radios(self, settings):
         test_form = CheckboxesSampleForm()
         test_form.helper = FormHelper()
@@ -256,9 +274,15 @@ class TestBootstrapLayoutObjects:
             assert html.count('<div class="panel-group"') == 1
             assert html.count('<div class="panel-heading">') == 2
         elif settings.CRISPY_TEMPLATE_PACK == "bootstrap4":
-            assert html.count('<div id="accordion"') == 1
+            match = re.search('div id="(accordion-\\d+)"', html)
+            assert match
+
+            accordion_id = match.group(1)
+
             assert html.count('<div class="card mb-2"') == 2
             assert html.count('<div class="card-header"') == 2
+
+            assert html.count('data-parent="#{}"'.format(accordion_id)) == 2
 
         assert html.count('<div id="one"') == 1
         assert html.count('<div id="two"') == 1