+++ CHANGELOG.md
@@ -1,9 +1,27 @@
 # CHANGELOG for django-crispy-forms
 
+## 1.11.2 (2021-03-16)
+* Added HTML parsing tools to help testing of template packs (#1128)
+* Fixed rendering of all widget attributes for file fields (#1130)
+
+See the [1.11.2 Milestone](https://github.com/django-crispy-forms/django-crispy-forms/milestone/15?closed=1) for the full change list.
+
+## 1.11.1 (2021-02-21)
+* Added `custom-select` for select widgets within prepended and appended fields (#1116)
+
+See the [1.11.1 Milestone](https://github.com/django-crispy-forms/django-crispy-forms/milestone/14?closed=1) for the full change list.
+
+## 1.11.0 (2021-01-30)
+* Implemeneted `custom-select` for Select widgets in the Bootstrap4 Template Pack (#1091)
+* Fixed `data-parent` in AccodrianGroup (#1091)
+* Documentation improvements
+
+See the [1.11.0 Milestone](https://github.com/django-crispy-forms/django-crispy-forms/milestone/13?closed=1) for the full change list
+
 ## 1.10.0 (2020-11-18)
 * Fixed test causing `SystemCheckError` in some cases. (#1075)
 * Radio and Checkbox now respect the field's `disabled` attribute when using the Bootstrap4 template pack. (#1057)
-* A number of documentation improvements. 
+* A number of documentation improvements.
 
 See the [1.10.0 Milestone](https://github.com/django-crispy-forms/django-crispy-forms/milestone/12?closed=1) for the full change list